Cont: 공유하기 기능 컴포넌트 분리
