+++ src/components/Kakao/KaKaoShare.tsx
@@ -1,5 +1,4 @@
 /** @jsxImportSource @emotion/react */
-import { css } from '@emotion/react';
 import { useEffect } from 'react';
 
 declare global {
